Added Integration Tests
