+++ integrations/nifi/nifi-nar-bundles/nifi-standard-services/nifi-spark-service-bundle/nifi-spark-service/src/test/java/com/thinkbiganalytics/nifi/v2/spark/SparkJobserverServiceTest.java
@@ -65,23 +65,6 @@ public class SparkJobserverServiceTest {
     private static final String syncTimeout = "600";
 
     /**
-     * Default Context Creation Properties
-     */
-    private static final String numExecutors = "1";
-    private static final String memPerNode = "512m";
-    private static final String numCPUCores = "2";
-    private static final SparkContextType sparkContextType = SparkContextType.SPARK_CONTEXT;
-    private static final int contextTimeout = 0;
-    private static boolean async = false;
-
-    /**
-     * Default Execute Spark Context Job Properties
-     */
-    private static final String appName = "test-app";
-    private static final String classPath = "spark.jobserver.WordCountExample";
-    private static final String args = "input.string = a b c a b see";
-
-    /**
      * Test runner
      */
     private final TestRunner runner = TestRunners.newTestRunner(new AbstractProcessor() {
@@ -135,134 +118,6 @@ public class SparkJobserverServiceTest {
         runner.enqueue(new byte[0]);
         results = ((MockProcessContext) runner.getProcessContext()).validate();
         Assert.assertEquals(0, results.size());
-    }
-
-    /**
-     * Verify creating Spark Context and then deleting the Context
-     */
-    @Test
-    public void testContextCreationAndDeletion() throws Exception {
-        // Test creating context
-        String contextName = "testContextCreationAndDeletion";
-
-        boolean created = sparkJobserverService.createContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async);
-        Assert.assertTrue(created);
-
-        boolean deleted = sparkJobserverService.deleteContext(contextName);
-        Assert.assertTrue(deleted);
-    }
-
-    /**
-     * Verify two threads attempting to create a Spark Context at once
-     */
-    @Test
-    public void testDuplicateContextCreation() throws Exception {
-        // Test creating context
-        String contextName = "testDuplicateContextCreation";
-
-        CreateSparkContext createSparkContext1 = new CreateSparkContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async, sparkJobserverService);
-        Thread thread1 = new Thread(createSparkContext1);
-
-        CreateSparkContext createSparkContext2 = new CreateSparkContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async, sparkJobserverService);
-        Thread thread2 = new Thread(createSparkContext2);
-
-        thread1.start();
-        thread2.start();
-
-        thread1.join();
-        thread2.join();
-
-        boolean contextExists = sparkJobserverService.checkIfContextExists(contextName);
-
-        Assert.assertTrue(contextExists);
-
-        sparkJobserverService.deleteContext(contextName);
-    }
-
-    /**
-     * Verify creating a Spark Context which times out and gets deleted
-     */
-    @Test
-    public void testContextTimeout() throws Exception {
-        // Test creating context
-        String contextName = "testContextTimeout";
-        int contextTimeout = 1;
-
-        boolean created = sparkJobserverService.createContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async);
-        Thread.sleep(15000);
-
-        boolean contextExists = sparkJobserverService.checkIfContextExists(contextName);
-        Assert.assertFalse(contextExists);
-    }
-
-    /**
-     * Verify creating two Spark Contexts
-     */
-    @Test
-    public void testMultipleContextCreation() throws Exception {
-        // Test creating context
-        String contextOneName = "testMultipleContextCreationOne";
-        String contextTwoName = "testMultipleContextCreationTwo";
-
-        CreateSparkContext createSparkContext1 = new CreateSparkContext(contextOneName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async, sparkJobserverService);
-        Thread thread1 = new Thread(createSparkContext1);
-
-        CreateSparkContext createSparkContext2 = new CreateSparkContext(contextTwoName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async, sparkJobserverService);
-        Thread thread2 = new Thread(createSparkContext2);
-
-        thread1.start();
-        thread2.start();
-
-        thread1.join();
-        thread2.join();
-
-        boolean contextOneExists = sparkJobserverService.checkIfContextExists(contextOneName);
-        boolean contextTwoExists = sparkJobserverService.checkIfContextExists(contextTwoName);
-
-        sparkJobserverService.deleteContext(contextOneName);
-        sparkJobserverService.deleteContext(contextTwoName);
-
-        Assert.assertTrue(contextOneExists && contextTwoExists);
-    }
-
-    /**
-     * Verify executing Spark Context Job
-     */
-    @Test
-    public void testExecuteSparkContextJob() throws Exception {
-        String contextName = "testExecuteSparkContextJob";
-
-        sparkJobserverService.createContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async);
-        boolean executed = sparkJobserverService.executeSparkContextJob(appName, classPath, contextName, args, async);
-        sparkJobserverService.deleteContext(contextName);
-
-        Assert.assertTrue(executed);
-    }
-
-    /**
-     * Verify executing Spark Context Jobs in Parallel
-     */
-    @Test
-    public void testParallelExecuteSparkContextJobs() throws Exception {
-        String contextName = "testParallelExecuteSparkContextJobs";
-
-        sparkJobserverService.createContext(contextName, numExecutors, memPerNode, numCPUCores, sparkContextType, contextTimeout, async);
-
-        ExecuteSparkContextJob executeSparkContextJob1 = new ExecuteSparkContextJob(appName, classPath, contextName, args, async, sparkJobserverService);
-        Thread thread1 = new Thread(executeSparkContextJob1);
-
-        ExecuteSparkContextJob executeSparkContextJob2 = new ExecuteSparkContextJob(appName, classPath, contextName, args, async, sparkJobserverService);
-        Thread thread2 = new Thread(executeSparkContextJob2);
-
-        thread1.start();
-        thread2.start();
-
-        thread1.join();
-        thread2.join();
-
-        sparkJobserverService.deleteContext(contextName);
-
-        Assert.assertTrue(executeSparkContextJob1.jobSuccessfull() && executeSparkContextJob2.jobSuccessfull());
     }
 
     /**
